% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_rex.R
\name{ask_rex}
\alias{ask_rex}
\title{Ask Rex}
\usage{
ask_rex(topic = "?")
}
\arguments{
\item{topic}{What you want to know about; either leave it blank for motivation or specify 'ttest', 'mean'... Input must be in quotes!}
}
\value{
Returns advice and explanations on R; sometimes Rex asks follow-up questions
}
\description{
This function lets you ask dino Rex for help.
You can try to ask for specific functions and topics or just use the function as-is
for some motivational words of encouragement. :)
}
\examples{
# Get motivation
ask_rex()

# Get specific information
ask_rex("mean")
ask_rex(topic = "ttest")

}
