% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiFiX_data_raw.R
\name{BiFiX_data_raw}
\alias{BiFiX_data_raw}
\title{BiFiX - Big Five raw data}
\format{
A data frame with 760 rows and 81 columns:
\describe{
\item{$validation_var}{validation variable}
\item{.wx.1.id}{ID}
\item{.wx.3.experimental_condition}{Experimental condition}
\item{big_...}{Big Five items (A - Agreeableness, C - Conscientiousness,
E - Extraversion, N - Neuroticism, O - Openness)}
...
}
}
\source{
\url{https://github.com/iscience-kn/BiFiX}
}
\description{
Raw data from the BiFiX study (submitted for publication in BRM)
}
