% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getmode}
\alias{getmode}
\title{Get the mode of a vector or variable}
\usage{
getmode(v)
}
\arguments{
\item{v}{a vector of variable in dataframe}
}
\value{
Returns the mode of v, i.e., the value occurring most often
}
\description{
The mode is the number that appears the most often in a vector or variable.
This measure is especially important for nominal variables since the mean
or median cannot meaningfully be reported.
}
\examples{

vec <- c("A", "A", "B")
getmode(vec)

}
