% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informed_consent.R
\name{informed_constent}
\alias{informed_constent}
\title{Informed Consent Filter}
\usage{
informed_constent(dataframe, varname = "informedconsent")
}
\arguments{
\item{dataframe}{A dataframe of data collected with WEXTOR}

\item{varname}{character. Variable name of informed consent variable. Defaults to "informedconsent".}
}
\value{
Returns data with cases filtered out that did not check the informed constent box
}
\description{
One of the best practices in data collection is including and informed consent question.
With this function you can automatically filter out rows where participants did not explicitly consent to their data being used.
}
