% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namepref.R
\name{namepref0}
\alias{namepref0}
\title{Add variable name prefix (without substitution)}
\usage{
namepref0(dataframe, pref_new)
}
\arguments{
\item{dataframe}{A dataframe}

\item{pref_new}{New prefix character to add to all variable namess in the dataframe have}
}
\value{
Returns dataframe with all variables renamed to start with the new prefix.
}
\description{
This function adds the new prefix and changes the dataframe's names to the new ones.
}
\examples{
bla <- tibble::tibble(x_ar = 1:5, y_ar = 6:10)
blo <- namepref(bla, "var")
names(bla)
names(blo)

}
