% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{path_to_file}
\alias{path_to_file}
\title{Get file path to \code{BiFiX_data_raw.csv} file}
\source{
This function is adapted from \code{palmerpenguins} (which is adapted from \code{readxl::readxl_example()}).
}
\usage{
path_to_file(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension;
\code{"BiFiX_data_raw.csv"} will work.
If \code{NULL}, the example file will be listed.}
}
\description{
This function makes the example data from the BiFiX study (submitted for publication in BRM) easy to access.
}
\examples{
path_to_file()
path_to_file("BiFiX_data_raw.csv")
head(read.csv(path_to_file("BiFiX_data_raw.csv")))
}
