% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removepref.R
\name{removepref}
\alias{removepref}
\title{Remove WEXTOR variable name prefixes (automatic)}
\usage{
removepref(dataframe)
}
\arguments{
\item{dataframe}{A dataframe, usefully containing variables that have the server prefix ".wx.#."}
}
\value{
Returns dataframe with those variables renamed that start with the old prefix.
}
\description{
This function strips the old prefix from the server variable names of the dataframe.
It identifies such variables that start with ".wx.#." where # stands for any number, or simply ".wx." and then removes
these prefixes for more easily legible and usable variable names in your data!
}
\examples{
bli <- tibble::tibble(.wx.1.ar = 1:5, y_ar = 6:10, .wx.z = 11:15)
blu <- removepref(bli)
names(bli)
names(blu)

}
