% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serious_check.R
\name{serious_check}
\alias{serious_check}
\title{Filter data frame by seriousness check}
\usage{
serious_check(dataframe, varname = "seriousness", keep = "participate")
}
\arguments{
\item{dataframe}{Data which contains a variable for the seriousness check.}

\item{varname}{Variable for the seriousness check. Defaults to \code{"seriousness"}.}

\item{keep}{Values from the variable to keep. Defaults to \code{"participate"}}
}
\value{
Filtered data
}
\description{
Filter your data frame by your item for the "seriousness check".
If the data comes from WEXTOR with no previous changes to the automatically created variable,
you will be able to accept all defaults in this function and simply input the data frame.
}
\examples{
# See vignette walkthrough

}
