% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausi_checks.R
\name{sess_length_check}
\alias{sess_length_check}
\title{Session Length Check}
\usage{
sess_length_check(dataframe, min_pages = 6)
}
\arguments{
\item{dataframe}{A dataframe containing WEXTOR data (needs to contain variables \code{page_trail} for the trail of
webpages in the study and \code{session_length} for the overall time each participant spent on the study)}

\item{min_pages}{Numeric. The minimum number of pages a participant must have
visited in the study for their participation to be considered plausible. Defaults to \code{6}.}
}
\value{
A dataframe with added variables \code{sess_length_clean} for only completed session lengths, otherwise NA
and \code{check_sess_length} for a flag of the most extreme 10\% of time needed to finish participation
}
\description{
This function takes a dataframe as input as well as the minimum number of webpages in your web experiment or study,
that a participant should have seen in order to be considered a complete (or plausible) participation.
It flags the most extreme 10\% of complete participations in the data as very slow or very fast.
}
\examples{

data <- read_WEXTOR(path_to_file("BiFiX_data_raw.csv"))
new_data <- sess_length_check(data)

}
