% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wob.R
\name{theme_wob}
\alias{theme_wob}
\title{Theme White on Black}
\usage{
theme_wob(max_font_size = 14, contrast = "high")
}
\arguments{
\item{max_font_size}{Maximum font size, used for the plot title (other font sizes are automatically
adjusted accordingly).}

\item{contrast}{Character to indicate the choice of color palette depending on the desired contrast.
Defaults to \code{"high"}. Should be one of \verb{"max", "high", "medium", "low", "pink" or "rex"}.}
}
\value{
Plot theme.
}
\description{
This theme option is built to be used with a \code{ggplot2} plot.
It offers several options for a dark theme, i.e. dark background with light writing.
}
\examples{
# Load the data
data <- read_WEXTOR("https://github.com/iscience-kn/BiFiX/raw/main/data/BiFiX_data.csv")

# Create any plot
ggplot(data, aes(age, color = gender, fill = gender)) + 
  geom_density(alpha = .5) + 
  theme_wob(contrast = "rex")

}
